+++ dnw_project/components/Hero.tsx
@@ -1,0 +1,30 @@
+import { motion } from 'framer-motion';
+import { FaPlayCircle } from "react-icons/fa";
+
+const Hero = () => (
+  <section className="h-screen">
+    <motion.div
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      transition={{ duration: 0.8 }}
+      className="flex flex-col items-center lg:flex-row justify-center gap-10"
+    >
+     <div className='mt-9 text-center px-8 flex flex-col max-w-lg'>
+        <img src="/d&w-logo.png" alt="hero" className='h-36'/>
+        <p className='mt-8'>Marvel Studios’ “Deadpool & Wolverine” delivers the ultimate, iconic, cinematic team-up throwdown. Come together in cinemas on July 24.</p>
+        <div>
+          <div className='mt-10 flex gap-10 justify-center'>
+            <button className='text-[#E81127] bg-[#FBDC0D] py-2 px-4 hover:bg-gradient-to-r hover:from-[#FFFFFF] hover:to-[#FBDC0D] font-semibold'>Get Tickets Now</button>
+            <button className='flex items-center gap-2 font-semibold'><FaPlayCircle /><span>Watch Trailer</span></button>  
+          </div>
+        </div>
+     </div>
+     <div>
+       <img src="/hero.png" alt="hero" className='h-96'/>
+     </div>
+    
+    </motion.div>
+  </section>
+);
+
+export default Hero;
